ApplyIPO/requirements.txt
asttokens==3.0.1
attrs==25.4.0
certifi==2026.2.25
decorator==5.2.1
executing==2.2.1
h11==0.16.0
idna==3.11
ipdb==0.13.13
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
outcome==1.3.0.post0
parso==0.8.6
pexpect==4.9.0
prompt-toolkit==3.0.52
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pysocks==1.7.1
python-dotenv==1.2.2
rich==14.3.3
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
traitlets==5.14.3
trio==0.33.0
trio-websocket==0.12.2
typing-extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
websocket-client==1.9.0
wsproto==1.3.2
